[Refactor/#178] 경매상품 response에 참여자수 추가

+++ src/main/java/com/yzgeneration/evc/domain/item/auctionitem/dto/AuctionItemResponse.java
@@ -60,10 +60,10 @@ public class AuctionItemResponse {
     @AllArgsConstructor
     public static class AuctionItemStatsResponse {
 
-        private int viewCount;
+        private Long viewCount;
 
-        private int likeCount;
+        private Long likeCount;
 
-        private int participantCount;
+        private Long participantCount;
     }
 }
